25 is nice. 26 needs a bit of work.
26 has an extra text field for SANs, but no functionality with the spreadsheet

Change the SAN field into a pop-out text-box that appears and populates relative to the count of the number of items being removed

The SAN to be inputted will be typed into an on-screen numpad, which is integrated into the pop-up window

This input will be inputted alongside the timestamp for this event

It is ok if the same timestamp and SAN number is used.

If the stock is removed from sheet 4.2_Items, then the timestamp info is inputted intp the sheet called 4.2_Timestamps.

If the stock is removed from sheet BR_Items, then the timestamp info is inputted intp the sheet called BR_Timestamps.