Modified the show_san_input function
Modified the show_san-input function so that the while True loop will continuously display the SAN input dialog until the user enters a valid SAN number that meets the length requirement. If the user enters an invalid SAN number (less than 5 characters), an error message is shown, and the dialog is displayed again.